fixes on entity_rules fix

+++ scripts/stack-transformer/train.sh
@@ -11,6 +11,9 @@ seed=$2
 # Load config
 . "$config"
 
+dir=$(dirname $0)
+parentdir="$(dirname "$dir")"
+
 # this is given by calling script to iterate over seeds
 checkpoints_dir="${CHECKPOINTS_DIR_ROOT}-seed${seed}/"
 
@@ -24,12 +27,12 @@ if [ -f "$ORACLE_FOLDER/train.rules.json" ];then
 fi
 
 # Copy entity_rules.json from oracle, created using train file
-if [ "$ENTITY_RULES" == "" ]; then
-    if [ ! -f "$ORACLE_FOLDER/entity_rules.json" ];then
-	cp $ORACLE_FOLDER/entity_rules.json $checkpoints_dir
-    fi
-else
+if [ -n "${ENTITY_RULES:-}" ] && [ -f "$ENTITY_RULES" ]; then
     cp $ENTITY_RULES $checkpoints_dir
+elif [ -f "$ORACLE_FOLDER/entity_rules.json" ];then
+    cp $ORACLE_FOLDER/entity_rules.json $checkpoints_dir
+else
+    python $parentdir/extract_rules.py $AMR_TRAIN_FILE $checkpoints_dir/entity_rules.json
 fi
 
 # Copy also dictionaries (we will need this for standalone)
